+++ packages/react/src/components/Dropdown/Dropdown.tsx
@@ -32,7 +32,7 @@ import {
 } from '../../lib'
 import List from '../List/List'
 import Ref from '../Ref/Ref'
-import DropdownItem from './DropdownItem'
+import DropdownItem, { DropdownItemProps } from './DropdownItem'
 import DropdownSelectedItem, { DropdownSelectedItemProps } from './DropdownSelectedItem'
 import DropdownSearchInput, { DropdownSearchInputProps } from './DropdownSearchInput'
 import Button from '../Button/Button'
@@ -512,16 +512,14 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
         variables,
         inputRef: this.inputRef,
       },
-      overrideProps: (predefinedProps: DropdownSearchInputProps) =>
-        this.handleSearchInputOverrides(
-          predefinedProps,
-          highlightedIndex,
-          rtl,
-          selectItemAtIndex,
-          toggleMenu,
-          accessibilityComboboxProps,
-          getInputProps,
-        ),
+      overrideProps: this.handleSearchInputOverrides(
+        highlightedIndex,
+        rtl,
+        selectItemAtIndex,
+        toggleMenu,
+        accessibilityComboboxProps,
+        getInputProps,
+      ),
     })
   }
 
@@ -598,7 +596,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
               key: (item as any).header,
             }),
         },
-        overrideProps: () => this.handleItemOverrides(item, index, getItemProps),
+        overrideProps: this.handleItemOverrides(item, index, getItemProps),
         render: renderItem,
       }),
     )
@@ -640,8 +638,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
               key: (item as any).header,
             }),
         },
-        overrideProps: (predefinedProps: DropdownSelectedItemProps) =>
-          this.handleSelectedItemOverrides(predefinedProps, item, rtl),
+        overrideProps: this.handleSelectedItemOverrides(item, rtl),
         render: renderSelectedItem,
       }),
     )
@@ -724,12 +721,20 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
     item: ShorthandValue,
     index: number,
     getItemProps: (options: GetItemPropsOptions<ShorthandValue>) => any,
-  ) => ({ accessibilityItemProps: getItemProps({ item, index }) })
+  ) => (predefinedProps: DropdownItemProps) => ({
+    accessibilityItemProps: getItemProps({
+      item,
+      index,
+      onClick: e => {
+        e.stopPropagation()
+        e.nativeEvent.stopImmediatePropagation()
+        _.invoke(predefinedProps, 'onClick', e, predefinedProps)
+      },
+    }),
+  })
 
-  private handleSelectedItemOverrides = (
+  private handleSelectedItemOverrides = (item: ShorthandValue, rtl: boolean) => (
     predefinedProps: DropdownSelectedItemProps,
-    item: ShorthandValue,
-    rtl: boolean,
   ) => ({
     onRemove: (e: React.SyntheticEvent, dropdownSelectedItemProps: DropdownSelectedItemProps) => {
       this.handleSelectedItemRemove(e, item, predefinedProps, dropdownSelectedItemProps)
@@ -747,7 +752,6 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
   })
 
   private handleSearchInputOverrides = (
-    predefinedProps: DropdownSearchInputProps,
     highlightedIndex: number,
     rtl: boolean,
     selectItemAtIndex: (
@@ -758,7 +762,7 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
     toggleMenu: () => void,
     accessibilityComboboxProps: Object,
     getInputProps: (options?: GetInputPropsOptions) => any,
-  ) => {
+  ) => (predefinedProps: DropdownSearchInputProps) => {
     const handleInputBlur = (
       e: React.SyntheticEvent,
       searchInputProps: DropdownSearchInputProps,
@@ -1027,7 +1031,6 @@ class Dropdown extends AutoControlledComponent<Extendable<DropdownProps>, Dropdo
     this.removeItemFromValue(item)
     this.tryFocusSearchInput()
     this.tryFocusTriggerButton()
-    e.stopPropagation()
     _.invoke(predefinedProps, 'onRemove', e, DropdownSelectedItemProps)
   }